Rename schemeshard_utils to schemeshard_index_utils and move non-index-related parts out of it
