+++ ydb/core/tx/schemeshard/schemeshard__operation_alter_table.cpp
@@ -1,7 +1,6 @@
 #include "schemeshard__operation_common.h"
 #include "schemeshard__operation_part.h"
 #include "schemeshard_impl.h"
-#include "schemeshard_utils.h"  // for TransactionTemplate
 
 #include <ydb/core/base/auth.h>
 #include <ydb/core/base/hive.h>
