+++ ydb/core/tx/schemeshard/schemeshard__operation_create_set_constraint_lock.cpp
@@ -1,6 +1,5 @@
 #include "schemeshard__operation_common.h"
 #include "schemeshard__operation_part.h"
-#include "schemeshard_utils.h"  // for TransactionTemplate
 
 #include <ydb/core/protos/flat_scheme_op.pb.h>
 #include <ydb/core/protos/flat_tx_scheme.pb.h>
