+++ ydb/core/tx/schemeshard/schemeshard__operation_drop_pq.cpp
@@ -2,7 +2,7 @@
 #include "schemeshard__operation_part.h"
 #include "schemeshard__operation.h"
 #include "schemeshard_impl.h"
-#include "schemeshard_utils.h"  // for PQGroupReserve
+#include "schemeshard_pq_helpers.h"  // for PQGroupReserve
 
 #include <ydb/core/base/subdomain.h>
 #include <ydb/core/persqueue/events/global.h>
