+++ ydb/core/tx/schemeshard/schemeshard__operation_move_index.cpp
@@ -2,7 +2,6 @@
 #include "schemeshard__operation_part.h"
 #include "schemeshard_impl.h"
 #include "schemeshard_path_element.h"
-#include "schemeshard_utils.h"  // for TransactionTemplate
 
 #include <ydb/core/base/path.h>
 #include <ydb/core/mind/hive/hive.h>
