+++ ydb/core/tx/schemeshard/ut_index/ut_fulltext_index.cpp
@@ -2,7 +2,7 @@
 #include <ydb/core/change_exchange/change_exchange.h>
 #include <ydb/core/scheme/scheme_tablecell.h>
 #include <ydb/core/testlib/tablet_helpers.h>
-#include <ydb/core/tx/schemeshard/schemeshard_utils.h>
+#include <ydb/core/tx/schemeshard/schemeshard_index_utils.h>
 #include <ydb/core/tx/schemeshard/ut_helpers/helpers.h>
